fix: `dropAllTables` should not raises exception when database is empty

+++ src/Dialects/Redshift.ts
@@ -61,6 +61,8 @@ export class RedshiftDialect implements DialectContract {
    */
   public async dropAllTables(schemas: string[]) {
     const tables = await this.getAllTables(schemas)
+    if (!tables.length) return
+
     await this.client.rawQuery(`DROP table ${tables.join(',')} CASCADE;`)
   }
